ARM: OMAP: AM33XX: Low power optimizations
As per the design team, in order to reduce VDDS_DDR
is suspend state the following things must be done:
1. Set the IOs to work in mDDR mode by setting mDDR_SEL
bit in the DDR_IO_CTRL register
2. Enable weak pull down for DQ and DM by configuring
the DDR_DATA0/1_IOCTRL registers
3. Disable VTP and also set P and N in VTP_CTRL to 1.

In the suspend state, A8 SRAM is configured to OFF state,
so to further reduce the power consumption SRAM LDO
is also configured to go to RET mode.

In the resume path all these registers need to restored
back to the pre-suspend state before any access to DDR
happens.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>