ARM: OMAP: AM33XX: CPSW: remove cpsw module warning fix
This patch fixes the warning generated while removing the cpsw module.
The warning is generated as net device is not unregistered before freeing
the net device and irq are not freed. Added freeing the irq and
unregister_netdev before free_netdev is called.

Also removes quirk as it is not required hence forth because all the
interrupts has to be used by the driver

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>