remoteproc/omap: set dependency on CONFIG_CMA
Commit f1ae98da8525c6b8b1c301c3a2b0bd2b6515cca2
(ARM: dma-mapping: remove unconditional dependency on CMA)
removes the unconditional selection of CMA for ARM platforms,
however omap remoteproc needs it for its pool allocation,
so make it depend on its Kconfig.

CMA is selected by the current omap2plus_defconfig and
arm_multiv7_defconfig, but the OMAP_REMOTEPROC configuration
should not rely on that default setting, before it can be
enabled.

Signed-off-by: Omar Ramirez Luna <omar.luna@linaro.org>
[s-anna@ti.com: revise commit log]
Signed-off-by: Suman Anna <s-anna@ti.com>