remoteproc: add safe version of vdev_to_rproc
vdev_to_rproc() assumes that the virtio device was created by remoteproc.
In some cases, however, this assumption might not hold. For instance,
when deducing a remoteproc from a given virtio device of a rpmsg
channel. In such cases the function should return an error instead of a
bogus pointer.

Signed-off-by: Ido Yariv <ido@wizery.com>
[s-anna@ti.com: move the rproc_type to internal header]
Signed-off-by: Suman Anna <s-anna@ti.com>