ARM: DRA7: hwmod data: add mmu data for ipu & dsp
A new MMU hwmod class and data structures are created to
represent the MMUs within the IPU and DSP processor subsystems
commonly present in DRA7xx. DRA7xx has two IPUs with an MMU
each, and a DSP with two MMUs, one for the processor core and
the other for the internal EDMA block.

NOTE:
1. The functional clocks for IPU1 and IPU2 are slightly
   different. IPU2 functional clock is sourced directly from
   dpll_core_h22x2_ck, while IPU1 has a mux clock for which one
   of the inputs is dpll_core_h22x2_ck. This mux clock is expected
   to be configured to be sourced from the dpll_core_h22x2_ck in
   turn, so that both IPU1 and IPU2 run at the same clock frequency.
2. The DSP processor subsystem in DRA7xx has two MMUs, one for the
   processor port and another for an EDMA port. Both these MMUs
   share a common reset line, the MMU on the EDMA port is expected
   to be mirror-programmed alongside the primary MMU, so the reset
   info is only added for the MMU associated with the processor
   port.

Signed-off-by: Suman Anna <s-anna@ti.com>