iommu/omap: add support for configuring dsp iommus on DRA7xx
The DSP MMUs on DRA7xx SoC requires configuring an additional
MMU_CONFIG register present in the DSP_SYSTEM address space.
This setting dictates whether the DSP Core's MDMA and EDMA
traffic is routed through the respective MMU or not. Add the
support to the OMAP iommu driver so that the traffic is not
bypassed when enabling the MMU0.

The MMU_CONFIG register has two different bits for enabling
each of these two MMUs present in a DSP processor sub-system
on DRA7xx. An id field is added to the OMAP iommu object to
identify and enable each IOMMU. A proper value is assigned to
this field only when the DSP_SYSTEM address space is present.
Logic is added to ensure that this is checked only for DSP
MMUs on DRA7xx SoCs and ignored for all others.

TODO:
1. Evaluate if the MMU_CONFIG register programming needs to
   be done in the probe and release functions instead so that
   the DSP traffic always goes through the corresponding MMUs,
   irrespective of whether the actual MMU itself is enabled or
   not. It is currently not done since probe does not do any
   dynamic programming.
2. This common register space logic can probably be implemented
   using a syscon device for the shared DSP_SYSTEM space, but
   would still need identifying the index of the MMU for
   programming the syscon register.

Signed-off-by: Suman Anna <s-anna@ti.com>