spi/qspi: use gpio to select qspi on AM43x epos evm
On AM43x epos evm, use of qspi and Nand is mutually exclusive. There is
a gpio present which is used to select between qspi and Nand.

As far as kernel is concerned, idea is to enable both qspi and nand
in dts, and able to use one of them based on user requirement.

The patch below makes use of the module parameter(enable_qspi) and expects
nand to use something similar(enable_nand) to get around this.

With this approach,
QSPI on J6: Works fine.
QSPI on AM43x-epos-evm "with" enable_qspi=1 in bootargs works fine
and qspi flash gets selected.
QSPI on AM43x-epos-evm "without" enable_qspi in bootargs prints a
message in the bootlog that "flash not recognised" as the flash does
not get selected.

This may not be the good solution, posting this to get some more feedbacks
on handling this out in a better way.

Signed-off-by: Sourav Poddar <sourav.poddar@ti.com>