remoteproc/omap: remove the unused fields from platform data
The following fields: .name, .oh_name, .oh_name_opt, .firmware, .ops
and .mbox_name, are removed from the platform data, as these are no
longer needed after the addition of DT support to the OMAP remoteproc
driver.

The .name field was used to give a name to the remoteproc, and this
can be replaced with the device name. The .ops field was never used
by the OMAP remoteproc driver. The .mbox_name is used to define the
sub-mailbox node used for communication with the remote processor,
and will be retrieved using the 'mboxes' property in the DT node. The
.firmware field is encoded directly in the OMAP remoteproc driver and
is retrieved using driver match data.

The .oh_name field is used to define the primary hwmod for the processor
node, and is represented using the 'ti,hwmods' property in the DT node.
The .oh_name_opt field was primarily defined to identify the hwmod for
the second cpu in a dual Cortex-M3/M4 IPU processor sub-system. This
hwmod entry is no longer defined usually, but rather a single hwmod
representing both the processors in the IPU sub-system is defined.
A single firmware image (either in SMP-mode or a combined image for
non-SMP mode) is used, with both the resets released together always
as part of the device management. Any power management and recovery
aspects require that both the processors be managed as one entity due
to the presence of shared MMU and unicache within the IPU sub-system.

The OMAP remoteproc platform data structure is eventually expected
to be removed completely once the other dependencies with the
mach-omap layer are met.

Signed-off-by: Suman Anna <s-anna@ti.com>