drm/tilcdc: adding more guards to prevent selecting invalid modes
The tilcdc has a number of limitations for the allowed sizes of
the various adjustable timing parameter.  Some modes are outside
of these timings.  This commit will prune modes that report timings
that will overflow the allowed sizes in the tilcdc.

Created by: Darren Etheridge <detheridge@ti.com>
Modified by Cody Lacey for the 3.8 kernel.