rpmsg: add api for creating and deleting rpmsg channels
The rpmsg channels are currently created and deleted only through
the device announcements from a remote processor. Two new exported
functions, rpmsg_create_channel and rpmsg_destroy_channel, are added
to be able to create and delete a rpmsg channel on a particular
virtual processor.

This is required for creating/deleting a channel from the HLOS-side
(needed by the rpmsg socket driver).

TODO:
The rpmsg socket driver should be fixed to not rely on this API going
forward, and instead leverage rpmsg_create_ept/rpmsg_delete_ept() API
so that the recovery state machine becomes less racy with userspace.

Signed-off-by: Robert Tivy <rtivy@ti.com>
[s-anna@ti.com: add delete channel api, documentation & checkpatch fixes]
Signed-off-by: Suman Anna <s-anna@ti.com>