ARM: omap: add DT support for deasserting hardware reset lines
Original author of this patch's code is:

    Matt Ranostay <mranostay@gmail.com>

This optional binding extension allows specification of a hwmod
and associate hardware reset line which should be deasserted for
the device to be functional.

The implementation works but is suboptimal. The problem is that this
deassertion occurs before clocks are enabled and we are warned that the
hard reset failed. Ideally the list of reset lines requested to be
deasserted would be cached and used within the hwmod enable sequencing
(instead of it just returning if any hardware reset line is asserted).

This reset line deassertion step is required to use the PRU-ICSS on
AM335x boards.