ASoC: simple-card: simplify code
[ Upstream commit b367a3252b365fd545fc589a40cad3976e73d7d7
45fce59496cbabd46761e6980c05c82d94d08eaa
ca65b492c7a265b220f763fd68bf87391213248f
520084729267ac8df1651ad2f118a1d4a631a10a
5ca8ba4180a6f629d51dba699b4a6428cc5eeba7
201a0eac7fe5e7a8fa33f0742304f885bc344d0d
2bee991460a838ee3c8064a6d880c4e7bc41717a
7722f830a45f7fbb8f2f7b23265793980bdf3397 ]

In the non-DT sequence, the platform data is copied as a whole to the
dynamic card info and the same pointer 'cinfo' is used to copy the
platform information to the card.

Use 'priv' as the pointer to the dynamic card info and copy only the
useful information from the platform data.

Signed-off-by: Jean-Francois Moine <moinejf@free.fr>
Signed-off-by: Mark Brown <broonie@linaro.org>
Signed-off-by: Darren Etheridge <detheridge@ti.com>