drm/tilcdc: rewrite pixel clock calculation
Updating the tilcdc DRM driver code to calculate the LCD controller
pixel clock more accurately.  Based on a suggested implementation by Tomi
Valkeinen.

The current code does not work correctly and produces wrong results with many
requested clock rates. It also oddly uses two different clocks, a display pll
clock and a divider clock (child of display pll), instead of just using the
clock coming to the lcdc.

This patch removes the use of the display pll clock, and rewrites the
code to calculate the clock rates. The idea is simply to request a clock
rate of pixelclock*2, as the LCD controller has an internal divider
which we set to 2.

Signed-off-by: Darren Etheridge <detheridge@ti.com>