PM / Voltagedomain: Add generic clk notifier handler for regulator based dynamic voltage scaling
This patch provides helper functions for drivers that wish to scale
voltage through the clock rate-change notifiers. The approach taken
is that the user-driver(cpufreq/devfreq) do not care about the
details of the OPP table, nor does it care about handling the voltage
regulator directly.

By using the clk notifier flags, we are able to sequence the operations
in the right order. The current logic is heavily influenced by
implementation done in cpufreq-cpu0.

Based on an original series by Mike Turquette <mturquette@linaro.org>

Signed-off-by: Nishanth Menon <nm@ti.com>