wl18xx: move to new firmware (wl18xx-fw-3.bin)
[ Upstream commit 028e7243ac411c3aba7a754bcc775c2fbb0b3e5c ]

This fw is not backward compatible with older
firmware (due to api changes), so use bump
the firmware name as well.

Some modifications were done to the driver-fw api
in order to support multiple APs.

Additionally, some of the consts (such as max stations,
max links and max RX BA sessions) were changed.

Signed-off-by: Arik Nemtsov <arik@wizery.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>