iommu/omap: change the attach detection logic
The OMAP IOMMU driver allows only a single device (eg: a rproc
device) to be attached per domain. The current attach detection
logic relies on a check for an attached iommu for the respective
client device. Change this logic to use a dedicated flag in
preparation for supporting multiple iommu devices to be bound to
a single iommu domain, and thereby to a client device.

Signed-off-by: Suman Anna <s-anna@ti.com>