ARM: OMAP2+: devices: use sentinel terminated iommu archdata for ISP
A client user of an IOMMU is tied to use a single IOMMU at the moment,
and both of them are bound together through the client user device's
archdata field. The archdata field is expected to be defined as a
single element currently.

Update the OMAP3 ISP device to use a sentinel terminated iommu archdata
array. This sentinel form allows the OMAP IOMMU driver core to be adapted
to recognize multiple IOMMUs without having to change the iommu archdata
structure. OMAP3 ISP will still use only one MMU device, but this form
allows the OMAP3 ISP driver to work properly both before and after the
OMAP IOMMU driver is enhanced to program multiple IOMMUs together.

Signed-off-by: Suman Anna <s-anna@ti.com>