omapdss: Add DRA74xx as an OMAPDSS version
DRA74x has a slightly modified version of the DSS on OMAP5. DRA74x has 3
instances of DPI, no DSI or RFBI. It has 2 VIDEO PLLs, which are very similar
to the DSI PLLs.

Create a new OMAPDSS version for it. Use omap5 structs for initializations in
dss_features, dss, dispc, pll and hdmi helper libraries.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>