soc: ti: Add wkup_m3_ipc driver
Introduce a wkup_m3_ipc driver to handle communication between the MPU
and Cortex M3 wkup_m3 present on am335x.

This driver is responsible for actually booting the wkup_m3_rproc and
also handling all IPC which is done using the IPC registers in the control
module, a mailbox, and a separate interrupt back from the wkup_m3. A small
API is exposed for executing specific power commands, which include
configuring for low power mode, request a transition to a low power mode,
and status info on a previous transition.

Also in this patch add platform data structures to be used to pass required
wkup_m3 APIs to platform code as that must remain built in at this time
which this driver supports loading as a module.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>