mfd: twl6040: Clear the interrupt ID register before requesting IRQ
[ Upstream commit 02d02728ea8bb33dee221d7dcfeb3022d7635927 ]

If for some reason the boot loader enabled the audpwron GPIO we will have
pending IRQs to be handled. This seams to break twl6040 for some reason
leading to non working i2c communication (i2c timeouts). Clearing the INTID
register after we requested the audpwron GPIO (and set it to low) will
ensure that the chip will operate normally in this case as well.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Darren Etheridge <detheridge@ti.com>