mfd: twl6040: Optional clk32k clock handling
[ Upstream commit 68bab8662f49b9e158f1d32f11becd4e48c04079 ]

In certain boards the source for the clk32k clock can be gated. In these
boards the clk32k clock can be provided to the driver and it is going to be
enabled/disabled when it is needed.
If the clk32k clock is not provided the driver will assume that it is always
running.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Darren Etheridge <detheridge@ti.com>