remoteproc: introduce watchdog and device exception error types
Introduce two new error types: watchdog and generic device exception,
that can be used for error reporting and recovery.

A Watchdog error is detected through a timer interrupt, due to that
timer not getting refreshed by a remote processor because it is stuck
in a busy loop. The generic device exception is a catch-all for all
other internal exceptions that can be reported to the host processor
through some mechanism.

Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>