drm/tilcdc add call to pm_runtime_irq_safe
set_scanout is called in irq context, which then calls pm_runtime_get_sync
which is not irq safe.  Adding a call to pm_runtime_irq_safe allows
pm_runtime_get_sync to be called from irq.  This eliminates a kernel oops from
occuring when modetest -v is used to do a flip test with pm enabled in the
kernel.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Reviewed-by: Felipe Balbi <balbi@ti.com>