remoteproc/pruss: define platform data for PRU remoteproc
The PRUSS has a PRCM reset line associated with the IP. The reset
needs to be programmed properly before accessing any of the internal
registers in the PRUSS. This functionality is achieved through the
omap_device layer, which is not exposed outside of mach-omap2 layer.
Define a platform data structure for PRUSS so that this API can be
invoked through platform data ops from the driver.

Signed-off-by: Suman Anna <s-anna@ti.com>