ARM: DRA7: hwmod: Add dra74x and dra72x specific ocp interface lists
To deal with IPs which are specific to dra74x and dra72x, maintain seperate
ocp interface lists, while keeping the common list for all common IPs.

For now move USB OTG SS4 to dra74x only list since its unavailable in
dra72x. The dra72x only list is empty for now and a placeholder for
future hwmod additions for CAL etc which are specific to dra72x.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>