remoteproc/core: introduce post loading handler
PRU custom resource handler for interrupts requires the event channel
information to be loaded into device memory, but the current core
code has the firmware segments loaded only after handling of the
loading handlers. Introduce a post loading handler, and process the
custom resource handler in this.

TODO: We may have to introduce the ability to process a custom resource
both before and after loading segments, as it really depends on what
is implemented within these handlers.

Signed-off-by: Suman Anna <s-anna@ti.com>