OMAPDSS: panel NEC-NL8048HL11 DT support
[ Upstream commit b36250d8b5a390ffc395c9c2cd5c39f722a276f4 ]

We don't have any working boards using this panel right now, and the
panel driver looks odd compared to the panel specs. For example, the
panel spec does not mention any QVGA pin.

So, while this patch adds DT support to the driver, it's not really
supported and there are not bindings documentation for the panel until
someone can verify how the panel actually works.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Cc: Erik Gilling <konkers@android.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>