remoteproc/core: add a rproc ops for performing address translation
The rproc_da_to_va API is currently used to perform any device to
kernel address translations to meet the different needs of the remoteproc
core/drivers (eg: loading). The functionality is achieved within the
remoteproc core, and is limited only for carveouts allocated within the
core or to internal memories published through the resource table.

A new rproc ops, da_to_va, is added to provide flexibility to platform
implementations to perform the address translation themselves when the
above conditions cannot be met by the implementations. The rproc_da_to_va()
is expanded to take in an additional flags field and invoke this ops if
present, and fallback to regular processing if the platform implementation
cannot provide the translation.

Signed-off-by: Suman Anna <s-anna@ti.com>