remoteproc/pruss: add pru-specific debugfs support
The remoteproc core driver creates certain standard debugfs
entries, that does not give a whole lot of useful information
for the PRUs. The PRU remoteproc driver is enhanced to add
additional debugfs entries for PRU. These will be auto-cleaned
up when the parent rproc debug directory is removed.

The added debugfs support adds two new entries: 'regs' and
'single_step'. The 'regs' dumps out the useful CTRL sub-module
registers as well as 32 registers each of the GPREGs and CT_REGs.
The GPREGs and CT_REGs though are printed only when the PRU is
halted and accessible as per the IP design.

The 'single_step' utilizes the single-step execution of the PRU
cores. Writing a non-zero value performs a single step, and a
zero value restores the PRU to execute in the same as mode as
before the first single step. (note: if the PRU is halted
because of a halt instruction, then no change occurs).

Signed-off-by: Suman Anna <s-anna@ti.com>