ARM: dts: AM4372: Add the PRU-ICSS1 DT node
Add the DT node for the PRU-ICSS1 on AM437x family of SoCs.
The DT node uses all standard properties. The regs property
defines the addresses for various modules within the PRUSS,
including all the Data RAMs, the Instruction RAM per core,
the INTC, CFG, Debug and Control sub-modules.

The two PRU cores within the PRUSS are represented as child
nodes (so platform devices themselves) of the PRUSS parent
node.

The PRU-ICSS1 on AM437x is very similar to the PRUSS in AM33xx,
except for variations in the RAM sizes and the number of
interrupts coming into the MPU INTC.

Signed-off-by: Suman Anna <s-anna@ti.com>