ARM: dts: AM4372: Add PRUSS1 PRU mboxes to mailbox node
The mailbox DT node data is updated with additional mbox data to
communicate with the PRU0 (User Id 1) and PRU1 (User Id 2) cores
within the PRU-ICSS1 subsystem. The following are the assignments
of mailbox fifos between MPU and the remote PRU cores:

            MPU Tx/PRU Rx   MPU Rx/PRU Tx
    PRU0    2               3
    PRU1    4               5

This assignment is identical to the assignment of mailboxes for
the PRU-ICSS subsystem in AM33xx.

Signed-off-by: Suman Anna <s-anna@ti.com>