ARM: OMAP2+: extend pruss pdata-quirks to AM4372
The omap_device API is needed to perform the reset management for
any IP instances with PRCM RSTCTRL registers (hard reset lines).
This API is limited to the mach-omap2 layer, and cannot be exposed
to drivers layer directly.

AM4372 requires the same functional reset management with PRUSS
as on AM33xx, so extend the AM33xx pruss pdata quirks to AM4372
as well. AM33xx relied on pdata quirks and platform data ops for
the PRUSS IP, to plumb the required omap_device API with the PRUSS
remoteproc driver using the pdata ops to achieve the required reset
functionality. This was implemented in this fashion as there is no
separate reset driver at the moment.

Signed-off-by: Suman Anna <s-anna@ti.com>