OMAPDSS: Wait for DSI PLL hsdividers in respective driver itself
Waiting for the PLL HSdivider to become active in dss.c makes dss.c code depend
on the type of PLL being used. It also requires us to extract the DSI platform
device which the PLL is a part of.

Simplify this by making the callers to the clk source selection function wait
for the HSdividers to become active. With this change, the functions
dss_select_dispc/lcd/dsi_clk_source just take care of setting the DSS bits for
switching the clock source.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>