OMAPDSS: DSI: Use PLL library
Make the DSI driver use the pll helper library to create PLL instances, and
register to those. We keep the existing dsi pll calculation stuff so that DPI
can still use it. This will be removed when DPI also stops using it.

There are some hacks which make sure that both the old DSI PLL code path and
the new PLL library work together. This is needed because DPI still uses dsi
functions to configure the DSI PLL. The old DSI PLL code will be removed once
DPI also uses the PLL library.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>