Revert "drm: Replace crtc fb with primary plane fb (v3)"
This reverts commit dfb048da1b46c7169d8c3b3360b2f2ba1e71b5b4.

This was ported from 3.15 into the original ti-linux-3.14 tree, but
unfortuntately it breaks the build of omapdrm (which is not enabled in the
default omap2plus_defconfig).  Pulling in all the dependent commits to make
this not break the build looks like a bunch of work, it could be done later if
deemed necessary.  For now lets revert it, and try and put omapdrm back into a
sane state.

Conflicts:
	drivers/gpu/drm/omapdrm/omap_crtc.c