OMAPDSS: DPI: Use VIDEO PLLs on DRA7x
Make the DPI driver use VIDEOx DPLLs for DRA74x. This involves adding an entry
in dpi_get_pll_data() for DRA74x, and calling a dss dpll function to configure
the DSS_PLL_CONTROL control module mux register to route the PLL output to the
desired overlay manager.

We currently fix the mapping of the PLLs with the LCD overlay managers. There
should ideally be a method by which the driver figures out a unused PLL, and
configure the mux registers pixel clock paths accordingly.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>