OMAPDSS: DISPC: Choose the PLL type based on a dispc feature
We need to retrieve the pll_data handle to calculate the hsdivider clock rate.
The pll can be either a VIDEO PLL on DRA7x platforms, or a DSI PLL on other
platforms.

Retrieve the right pll_data handle using a dispc feat which choses the dss dpll
function to get pll_data for DRA7x, and the dsi pll function to get pll_data
for others.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>