HACK: drm/tilcdc: add vsync callback for use in omaplfb for gpu
Add a vsync callback registration API that is identical to what was done
for da8xx-fb.c.

Need to find if there is a better way using the DRM infrastructure from
kernel space.  Either that or change the userspace window manager stuff in
the gpu libraries to make use of the DRM provided syncronization
mechanisms.  There is a generic vblank_wait function proposal circulating the
dri-devel list at the moment that could replace this.

Signed-off-by: Darren Etheridge <detheridge@ti.com>