iommu/omap: fix iommu archdata name for DT-based devices
A device is tied to an iommu through its archdata field. The archdata
is allocated on the fly for DT-based devices automatically through the
.add_device iommu ops. The current logic assigns the name of the IOMMU
user device, instead of the name of the IOMMU device as required by
the attach logic. Fix this issue so that DT-based devices can attach
successfully to an IOMMU domain.

Signed-off-by: Suman Anna <s-anna@ti.com>