ARM: dts: DRA7: Add vtt regulator support
DRA7 evm REV G and later boards uses a vtt regulator for DDR3 termination
and this is controlled by gpio7_11. This gpio is configured in boot loader.
gpio7_11 is available only on Pad A22 on all boards, looking at the previous
board schematics pad A22 is connected only to EXP_P3. This pad in EXP_P3 is
currently not used.
Current code is configuring the pad A22 to its default function
spi2_cs1(muxmode 0), and as of now there are no users of the channel.
So removing the mux node for mcspi1 and marking it for gpio7_11 and adding
regulator node.
Since gpio7_11 is used to control VTT and should not be reset or kept in
idle state during boot up else VTT will be disconnected and ddr gets
corrupted. So marking gpio7 as no-reset and no-idle on init.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>