HACK: drm/fb_helper: enable panning support
Increase the size of the buffer that is created in the fbdev emulation
helpers.  And fill in the var structure with the amount that was allocated.

This should be parameterized so each DRM driver could pick its required
number of buffers allocated for /dev/fb0, however I don't see an easy way
to do this.  For now it is hardcoded and will effect all platforms that use
fb_cma_helper (only tilcdc for ti-linux-3.14.y).

Signed-off-by: Darren Etheridge <detheridge@ti.com>