ARM: AM33XX: Don't rewrite SDRAM_CONFIG register for DDR2 systems
When the EMIF (memory controller) resets after losing power, the
EMIF_SDRAM_CONFIG register gets loaded with the contents of the
CONTROL_SECURE_SDRAM_CONFIG register. This is so that reinitializations
of the SDRAM that occur when writing to this register are not necessary
at resume time.

Program the CONTROL_SECURE_SDRAM_CONFIG register once at init time rather
than the SDRAM_CONFIG register at every resume.

This is needed to make the most gains out of CPU idle, which is very
sensitive to return to active timing.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>