mmc: omap_hsmmc: enhance pinctrl support
Update omap_hsmmc driver to set the state of the pins to:
- "default" on resume
- "sleep" on suspend

This should make it possible to optimize energy usage for the pins both
for the suspend/resume cycle, and for runtime cases in-between mmc
transfers.

By optionally putting the pins into sleep state in the suspend callback
we can accomplish two things.
- minimize current leakage from pins and thus save power,
- prevent the IP from driving pins output in an uncontrolled manner,
which may happen if the power domain drops the domain regulator.

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>