ARM: OMAP2+: AM43XX: Remove flags from gpmc hwmod
Without initial soft reset of GPMC during hwmod setup the CLKCTRL
for GPMC will not transition to idle state which prevents PER_PD
transition so this patch allows it.

Also remove the HWMOD_INIT_NO_IDLE flags as it should be handled by
a dt property, which will be added in a later patch.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>