ARM: OMAP2+: omap_hwmod: Introduce HWMOD_NEEDS_REIDLE
Some hwmods will not properly assert signals to the PRCM after a
context loss if no driver is present which leads to issues with suspend.
omap_hwmod will idle them properly after boot but after the first context
loss they will be in the wrong state so suspend will no longer work as
there is no driver associated with the hwmod.

Introduce a new flag, HWMOD_NEEDS_REIDLE, to allow these modules to be
tracked and properly handled. omap_hwmod maintains a list of these
modules and uses a PM notifier to enable and then idle the hwmods
immediately after resume. omap_device will remove hwmods from this
list when a driver is bound and add the hwmods back if the driver is
removed to avoid any conflicts and allow the proper pm layer to handle
things when a driver is present.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>