ARM: AM43xx EVM: static configuration for LCD/HDMI switch (HACK)
AM43xx EVM boards have a gpio which switches between LCD and HDMI, and
additionally between analog audio and HDMI audio. As we don't have means
to handle that switch dynamically, we'll have separate .dts files for
LCD and HDMI configurations. We still need to set the switch GPIO value
to high or low, but we don't have means for that either.

This patch is a temporary solution, which adds pdata_quirks for the
EVMs. If a HDMI encoder node is found, the GPIO is set to HDMI mode.
Otherwise the GPIO is set to LCD mode.

[detheridge - fixed up to merge against ti-linux-3.14.y]

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Acked-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>