ARM: OMAP2+: AM43XX: add AM43xx support to AM33xx PM Code
Because am33xx and am43xx families are very similar we can reuse much
of the code needed for suspend resume. Add code for SCU control and
additional arguments to be passed to low-level asm code.

Also introduce am33xx_pm_ops to allow for operations specific to
each SoC to be executed at init, pre-suspend, and post-suspend that
can be set at init time.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>