remoteproc/omap: add support for DT nodes
OMAP4+ SoCs support device tree boot only. The OMAP remoteproc
driver is enhanced to support remoteproc devices created through
Device Tree, support for legacy platform devices has been
deprecated. The current DT support handles the IPU and DSP
processor subsystems on OMAP4 and OMAP5 SoCs.

The OMAP remoteproc driver expects the firmware images to have
fixed names. This used to be defined through platform data
previously, and are now coded into the driver. The following
names (same as prior kernel non-DT names) are to be expected
of the firmwares,
	OMAP4/OMAP5 IPUs: ducati-m3-core0.xem3
	OMAP5/OMAP5 DSPs: tesla-dsp.xe64T

The OMAP remoteproc driver relies on the omap_device, control
module and omap_hwmod layers for performing clock, reset and
boot vector management of the devices, but these are limited
only to the machine-specific layers in arch/arm. This management
is to be achieved through OMAP rproc specific platform data ops,
and the required implementations to boot and shutdown have been
added in the machine layer. These need to be plugged in to the
remoteproc devices through pdata quirks, for properly booting
the remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>