ARM: OMAP2+: Extend OMAP4 IPU/DSP pdata-quirks to OMAP5
OMAP5 requires the same device management pdata-quirks as OMAP4
for the IPU and DSP processor subsystems, so extend the OMAP4 IPU
and DSP pdata quirks to OMAP5 as well.

The pdata quirks do not use a starting address in the auxdata, as
there are only single instances of DSP and IPU on OMAP5, and to
also match with the current DT node definitions. The pdata quirks
will be removed once the reset dependencies against omap_device
& omap_hwmod layers are resolved.

Signed-off-by: Suman Anna <s-anna@ti.com>