CLK: TI: OMAP5: Add timer_sys_ck clock aliases for Timers
The OMAP DMTimer API, omap_dm_timer_set_source, can set the parent
of a timer node using 3 different values that use fixed parent names
for the clocks. The parent name, timer_sys_ck, is used for setting the
parent when used with the source index OMAP_TIMER_SRC_SYS_CLK. This
source index is used by the OMAP remoteproc drivers to set the parent
of the timers used to provide the remote processor's OS system clock,
and the parent name timer_sys_ck is not defined for OMAP5. Add these
aliases so that the sources for these timers can be set properly.

Signed-off-by: Suman Anna <s-anna@ti.com>