ARM: OMAP2+: Add watchdog timer ops to rproc pdata quirks
The OMAP remoteproc driver achieves watchdog functionality by
registering and handling a timer interrupt, which is triggered
upon a watchdog condition on the remote processor. The OMAP
DMTimer API though is limited to only the arch/arm/ layers,
so the API needs to be accessed through OMAP rproc platform
data ops.

Add the ops for retrieving a DMTimer IRQ id, and for clearing
the Timer IRQ status (minimum ops required for watchdog
functionality in addition to other DMTimer ops) to the common
timer ops for both the IPU and DSP platform devices on OMAP4
and OMAP5 SoCs.

Signed-off-by: Suman Anna <s-anna@ti.com>