remoteproc/omap: add support for DRA7xx remote processors
DRA7xx SoC has two IPU and up to two DSP processor subsystems for
offloading different computation algorithms. The IPU processor
subsystem contains dual-core ARM Cortex-M4 processors, and is
very similar to those on OMAP5. The DSP processor subsystem is
based on the TI's standard TMS320C66x DSP CorePac core.

Support has been added to the OMAP remoteproc driver through new
DRA7xx specific compatibles for properly probing and booting the
all the different processor subsystem instances on DRA7 - IPU1,
IPU2, DSP1 & DSP2. A build dependency with SOC_DRA7XX is added to
enable the driver to be built in DRA7xx-only configuration.

The firmware lookup logic has to be adjusted for DRA7xx as there
are (can be) more than one instance of both the IPU and DSP remote
processors for the first time in OMAP4+ SoCs. The names (same as
prior kernel non-DT names) for the firmware images are fixed for
each processor and is expected to be as follows,
	IPU1: dra7-ipu1-fw.xem4
	IPU2: dra7-ipu2-fw.xem4
	DSP1: dra7-dsp1-fw.xe66
	DSP2: dra7-dsp2-fw.xe66

Signed-off-by: Suman Anna <s-anna@ti.com>