ARM: OMAP2+: Add rproc pdata-quirks for IPUs & DSP1 on DRA7
The remote processors on DRA7xx requires similar device management
pdata-quirks (reset control, clocking, boot vector programming,
dmtimer API wrappers for enabling both system and watchdog timers)
as the IPU and DSP processors on OMAP4/OMAP5. All these pdata ops
are identical to the ones used on OMAP4/OMAP5 except for the boot
vector programming on DSPs on DRA7xx.

The pdata quirks for the most common remote processor subsystems
on DRA7xx - IPU1, IPU2 and the DSP1, have been added. The quirks
need to match the starting address in the auxdata for the respective
processors, as the same compatible string is used for all the
instances of the remote processor of the same type. The pdata
quirks will be removed once the dependencies against arch/arm
layers are resolved.

Signed-off-by: Suman Anna <s-anna@ti.com>